+++ tests/test-udp.c
@@ -53,7 +53,7 @@ static void _echo_server(void* arg) {
     char     buf[64];
     char     sender_host[46];
     uint16_t sender_port = 0;
-    int64_t  n = xylem_udp_recv(
+    int  n = xylem_udp_recv(
         udp, buf, sizeof(buf), sender_host, sizeof(sender_host), &sender_port);
     ASSERT(n == 5);
     ASSERT(memcmp(buf, "hello", 5) == 0);
@@ -75,7 +75,7 @@ static void _echo_client(void* arg) {
     ASSERT(xylem_udp_send(udp, "hello", 5, NULL, 0) == 0);
 
     char    buf[64];
-    int64_t n = xylem_udp_recv(udp, buf, sizeof(buf), NULL, 0, NULL);
+    int n = xylem_udp_recv(udp, buf, sizeof(buf), NULL, 0, NULL);
     ASSERT(n == 5);
     ASSERT(memcmp(buf, "world", 5) == 0);
 
@@ -127,7 +127,7 @@ static void _addr_receiver(void* arg) {
     char     buf[64];
     char     host[46];
     uint16_t port = 0;
-    int64_t  n = xylem_udp_recv(udp, buf, sizeof(buf), host, sizeof(host), &port);
+    int  n = xylem_udp_recv(udp, buf, sizeof(buf), host, sizeof(host), &port);
     ASSERT(n == 4);
     ASSERT(port == ctx->port_b);
 
@@ -169,7 +169,7 @@ static void _timeout_coro(void* arg) {
     xylem_udp_set_read_deadline(udp, deadline);
 
     char    buf[64];
-    int64_t n = xylem_udp_recv(udp, buf, sizeof(buf), NULL, 0, NULL);
+    int n = xylem_udp_recv(udp, buf, sizeof(buf), NULL, 0, NULL);
     ASSERT(n == -1);
 
     xylem_udp_close(udp);
@@ -204,7 +204,7 @@ static void _close_recv_coro(void* arg) {
     xylem_channel_send(ctx->ready, udp);
 
     char    buf[64];
-    int64_t n = xylem_udp_recv(udp, buf, sizeof(buf), NULL, 0, NULL);
+    int n = xylem_udp_recv(udp, buf, sizeof(buf), NULL, 0, NULL);
     ASSERT(n == -1);
 
     xylem_waitgroup_done(ctx->wg);
